Hotfix: css 수정 및 오류 해결

+++ src/Components/Feed/FeedDetail/index.tsx
@@ -33,6 +33,7 @@ export const FeedDetail = ({
   feedLikeData: LikeType[] | undefined;
 }) => {
   const [isLike, setIsLike] = useState(false);
+  const [deleteFeedId, setDeleteFeedId] = useState<string | null>(null);
 
   let isLogin = useAppSelector(state => state.login.isLogin);
   let loginUserId = useAppSelector(state => state.user.id);
@@ -146,6 +147,7 @@ export const FeedDetail = ({
 
   // 게시물 삭제
   const deleteFeedAlert = () => {
+    feedId && setDeleteFeedId(feedId);
     dispatch(
       alertActions.setModal({
         isModalOpen: true,
@@ -167,11 +169,12 @@ export const FeedDetail = ({
   });
 
   useEffect(() => {
-    if (isDelete) {
+    if (isDelete && deleteFeedId !== null) {
       deleteMutate(feedId);
       dispatch(alertActions.setIsClickOk());
+      setDeleteFeedId(null);
     }
-  }, [isDelete, feedId, dispatch, deleteMutate]);
+  }, [isDelete, deleteFeedId, dispatch, deleteMutate]);
 
   useEffect(() => {
     if (isError) {
@@ -223,10 +226,12 @@ export const FeedDetail = ({
           <h1 className="text-xl font-bold">{feedDetailData?.title}</h1>
         </div>
         <div className="flex justify-center items-center flex-col w-full mt-4 gap-4">
-          <div className="flex justify-between items-center w-full md:mt-8">
-            <div className="flex justify-between items-center w-full">
-              <div className="text-sm text-buttongray">
-                {createDate} 작성 | {updateDate} 편집
+          <div className="flex flex-col justify-between items-center w-full md:mt-8">
+            <div className="flex md:flex-row flex-col md:justify-between justify-center items-start w-full">
+              <div className="flex md:flex-row flex-col text-sm text-buttongray">
+                <span>{createDate} 작성 </span>
+                <span className="md:inline hidden mx-4">|</span>
+                <span>{updateDate} 편집</span>
               </div>
               <div className="flex align-center gap-8">
                 <Link to={`/channel/${feedDetailData?.user.id}?type=review`}>
